+++ src/CoordinateOperation/Params/EPSG10619.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * PHPCoord.
+ *
+ * @author Doug Wright
+ */
+declare(strict_types=1);
+
+/** @internal */ return [
+    'offsetsFile' => 'PHPCoord\\CoordinateOperation\\BYNCGVD28CGG2013a2010Provider',
+];