+++ src/CoordinateOperation/Params/EPSG10677.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * PHPCoord.
+ *
+ * @author Doug Wright
+ */
+declare(strict_types=1);
+
+/** @internal */ return [
+    'offsetsFile' => 'PHPCoord\\CoordinateOperation\\NTv2PD83ETRS89ThuringiaProvider',
+];